Add files via upload
Dylan Lucas - Created code for Login. When user tries to login, it checks the username and password and makes sure its in the database. It also makes sures that the password is associated with that username and not with another username. So login now works.

Roy Hampton - Created account creation password verification. Made sure Password is a strong password containing one uppercase, one lowercase, one special character, and is at least 8 charachters long using regex.