Add files via upload
Dylan Lucas - Added a menu for when, right now at least until more 'Employees' are added to the database, when Admin logs in they go to the Employee menu. This is the menu for where you enter in the Refund Number or Repair Number of a Customer and can update the status of any of there repairs or refunds. 

Also added the directory to the database tables, Television, Consoles, Computers, and Tablets, into the main application. User clicks Television then the television directory table will show up and so on and so forth for the others. So users can now see the database when logged in.