Add files via upload
Dylan Lucas - Added verification that none of the account creation text boxes are empty when a user tries to create an account. Also added feature of textbox turns Red with white text when its not completely correctly and turns back to white with black text when its correct so user knows what to fix at all times.