Dylan Lucas - Got rid of the structures and added a Cart table to the GSEDatabase.
Added a way to call and receive the variable for the username when somebody creates a new account and is instantly logged in or when they log in that way the user does not have to repeatedly type in there username.

Added function for the user to be able to add items in each directory, Television, Console, Tablet, Computer to there cart

Roy Hampton & Dylan Lucas - Created way for user to go to the Cart and see there current cart. Made sure it only checks there cart and not the entire database.

Currently unable to figure out how to update / refresh the datagridview for the cart, so in place of that for now, we put a Messagebox that lets the user know to see an updated cart they must restart the entire client.