Dylan Lucas - Set up the Receipt screen to where it will now send the user an email (corresponding with the email in the database associated with the account's username that they signed in with) and will give them the total price of there order which includes shipping, warranty price, and bundle price. Will tell whether the order contains a console, computer, television, or tablet, and gives the user there order number in the email so they can further track the package.
Also added the track order function on the main screen. User can only go to that screen if there is an order that exists, if no order exists user cannot go to screen.

Added same functionality for Go to Cart, no point in going to cart if cart has nothing in it.

Added same functionality to the Purchase Order. No point in allowing the user to progress if they have nothing in there cart to purchase.