Add files via upload
Dylan Lucas - Fixed a lot of issues with database stuff and stuff with the going to cart, removing item, and purchase order stuff as it was not working correctly because of the connection string. Made a database class to perform the connections and future SQL Statements. Hashed and Salted user passwords when they create there accounts. Stored CC information but Hashed and Salted the Credit Card Number (not sure if I am going to keep that way but currently it is that way.